refactor(tests): split test_chat_servicer.py, encoder_test.go, cost_test.go into focused modules (#203)
test_chat_servicer.py (763 lines) → test_extract_chat_reply.py,
test_send_chat_message.py (each ≤ 450 lines).

encoder_test.go (654 lines) → encoder_helpers_test.go,
encoder_schema_test.go, encoder_resilience_test.go (each ≤ 370 lines).

cost_test.go (663 lines) → cost_estimate_test.go, cost_counter_test.go,
cost_budget_test.go, cost_config_test.go (each ≤ 245 lines).

Removes all three from GRANDFATHERED_FILES in scripts/checks/file_size.py
so the CI file-size guard now enforces the 500-line limit on these modules.
