+++ FILEMAP.md
@@ -1,16 +1,16 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **439** tracked files
+> **445** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 121 |
+| `internal/` | 126 |
 | `docs/` | 81 |
 | `agents/` | 77 |
-| `tests/` | 70 |
+| `tests/` | 71 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
@@ -26,8 +26,8 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 158 |
-| `.go` | 118 |
+| `.py` | 159 |
+| `.go` | 123 |
 | `.md` | 97 |
 | `.yaml` | 25 |
 | `.rs` | 9 |
@@ -312,7 +312,10 @@
 │   │   ├── cache_test.go
 │   │   ├── config.go
 │   │   ├── cost.go
-│   │   ├── cost_test.go
+│   │   ├── cost_budget_test.go
+│   │   ├── cost_config_test.go
+│   │   ├── cost_counter_test.go
+│   │   ├── cost_estimate_test.go
 │   │   ├── reporter.go
 │   │   └── reporter_test.go
 │   ├── defaults/
@@ -376,7 +379,9 @@
 │   │       ├── context.go
 │   │       ├── context_test.go
 │   │       ├── encoder.go
-│   │       ├── encoder_test.go
+│   │       ├── encoder_helpers_test.go
+│   │       ├── encoder_resilience_test.go
+│   │       ├── encoder_schema_test.go
 │   │       ├── options.go
 │   │       └── stderr.go
 │   ├── planner/
@@ -512,7 +517,6 @@
 │           ├── test_base_agent.py
 │           ├── test_base_handle.py
 │           ├── test_builtin_tools.py
-│           ├── test_chat_servicer.py
 │           ├── test_defaults.py
 │           ├── test_dispatch_execute_actions.py
 │           ├── test_dispatch_lifecycle.py
@@ -524,6 +528,7 @@
 │           ├── test_episodic_memory_retention.py
 │           ├── test_episodic_memory_validation.py
 │           ├── test_event_dispatcher.py
+│           ├── test_extract_chat_reply.py
 │           ├── test_llm_client.py
 │           ├── test_llm_persona_agent.py
 │           ├── test_memory_injection.py
@@ -543,6 +548,7 @@
 │           ├── test_relationship_memory.py
 │           ├── test_relationship_memory_user.py
 │           ├── test_sandbox.py
+│           ├── test_send_chat_message.py
 │           ├── test_server_channel.py
 │           ├── test_server_lifecycle.py
 │           ├── test_server_load_agent.py
